requirements.txt
coverage==4.5.4
django==2.2.10
django-nose==1.4.6
djangorestframework==3.10.3
nose==1.3.7
